Remove duplicate poison tests and expand coverage
